Delete files using fs

+++ shopping-app-expressjs-mongodb/controllers/mongoose/admin.js
@@ -1,5 +1,6 @@
 const Product = require("../../models/mongoose/product");
 const { validationResult } = require("express-validator/check");
+const fileHelper = require("../../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -173,6 +174,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = price;
 
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
 
@@ -188,7 +190,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
   let productId = req.body.productId;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
+  Product.findById(productId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found!"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
     .then((result) => res.redirect("/admin/products"))
     .catch((err) => {
       const error = new Error(err);
